[IMP] mail: Composer check duplicates
## Issue

Only duplicate emails used to be checked when sending a mass mail.

However it is possible (e.g. using templates) to send a mass mail
to the same person containing different information.

## Solution

The existing functions to allow models to specify emails
processed in the past by some other means are kept.

A new check is added in the processing that checks the full contents
of the message, subject and attachment ids.

## Technical details
The strings are not hashed as most situations are:
- Sending the exact same mail to everyone
  -> Only need to check against one message
   -> Same complexity as hashing

- Sending all different emails
  -> Checking inequality of str is usually very fast

We also only check the actual attachment id, not its content.
As there is no situation were we would generate the same attachment
and not expect it to be sent twice.

## Removed logic

We can thus remove the 'document based' information
as it is implicitly infered from this check.

Task-2826811