[IMP] *: add slug and remove base url from all website content URLs
*: website_event_exhibitor, website_event_meet, website_hr_recruitment,
   website_slides

This follows the merge of [1].

Model implementing the _compute_website_url method sometimes added the
"base url" part which is not necessary (there is no case where we want
a website to display a record from another website and would thus need
an absolute URL to make the switch**) and sometimes forgot to add the
"slug" part.

**: Notice that the main point of this change is that the website_url
field of some of those models are now shown in the "Site" menu of the
website app. In that case, when clicking on a record, the user is
redirected to the iframe preview, with the right website and domain
without the need for it to be part of the object's URL.

[1]: https://github.com/odoo/odoo/commit/da3f4c2aff92cca23ca2d7a87fa0eb5b1d7ce5fe