[IMP] project: add the manual aal to project profitabilty, other generic upgrade
This commit purpose is
- to add the all the user adds manually to the
panel of project profitability if there is only project linked to the
analytic account.
- to remove the stage tracking for task dependencies, and instead log a
note to warn the user when all the blocking tasks are closed. If one of
those task is reopened, message is sent to warn that the task is
blocked again
- to display the name of the current task/ticket/project on the portal
page instead of the generic My tasks/etc
- to update the name of notification sub-types task/project dependcy

task-2960753