[IMP] mail: support scheduled date from template on composer
RATIONALE

Improve usage of composer in comment mode: support batch-posting, support
more configuration from templates, improve model.

SPECIFICATIONS

Purpose of this task is to correctly support scheduled_date from template on
both comment and mass mode in the composer.

It is currently mainly supported at template level, when using a template
to directly send emails. In this commit we add a field on composer that
takes the value from the template, and propagate it to the mail or messages
created when validating it.

As most template fields it can contain inline template code to be rendered
dynamically on target records, hence using a char field. Its rendering
is done using template that calls _parse_scheduled_datetime. It allows
to have an UTC and timezone agnostic value.

Task-2993872 (Mail: Support scheduled date in all composer flows)