[IMP] *: use many2one for language field instead of selection
- Replace selectiong field for language by a many2one.
- Remove lang counter field

The old lang field becomes mainly a related field of the new
many2one field. Except for the partner and translation model
as a lot of logic depends on the context key "lang" which is
a language code.
It's indeed easier to work with the code for translations or
even the website language instead of an id.
For these cases the lang field becomes a stored and computed field.
As it's possible to set the lang code in these models, we can't use
a related field because the lang code cannot be modified.

task-2856992