[IMP] mail: improve internal and external link behavior
PURPOSE:

We want to differentiate the behavior of internal links to Odoo documents and
the external links to any web page in order to allow the user to navigate
smoothly through Odoo from Discuss.

SPECIFICATION:

When a user clicks on an internal link in Discuss to an Odoo document:
- The document is open in the same tab in the browser
- The chat window opens with the conversation where the link was shared.
When a user clicks on an external link it should open in new tab.

The same behavior should apply to the link shared in Chatter/Activity Log Note.

Task-2636465