[IMP] base: separator and date time preview
when the user changes the date time preview or separator for particular the
language they have to go to the SO, Invoice, etc. to check the real value for
the format and if the value is not proper they have to change the format again.
after this commit, the user can see the preview just beside the format
to see if the given format is proper or not

task- 2622018