[IMP] stock*: added "Order to max" feature in replenishment
Before this commit
==================
There is no functionality to trigger replenishment before min qty is reached.

After this commit
=================
Button added in the list view of replenishment "Order to max". This button will
trigger replenishment even if min qty is not reached for each selected line if
Max Quantity - forecast > 0.

TaskID: 2964309