[WIP / PY] base: allow to group by properties
Purpose
=======
Allow to group record by their property values.

Technical
=========
The "group by" work like standard fields, with one difference for the
tags and the many2many properties. There's no "None" column for them.

The reason for that is that we need to check the many2many existence
in SQL (a value should be in the "None" column if it has an empty
array in database, or an array with only non-existing ids, if it has
at least 1 valid record id it shouldn't be in the "None" column). It
can't be done in python otherwise the count value might be wrong...
In SQL, it would require a "NOT IN (SELECT id FROM xxx)" which is very
inefficient, and so for simplicity we do not show the "None" column for
the many2many. For the tags it can be possible, but to make the
behavior consistent with the many2many, it hasn't been done.

The order in argument might not be always respected. If we order and
group by a property, the "falsy" domain (e.g. a record has been deleted)
will be merged (with an OR). If there's a "False" value in database,
the order will be respected even after the merge. But if no "False"
value are in database (so only non existing record ids, only removed
selection options, etc), the "Falsy" domain will be added at the end
(which might not respect the order received in argument for that
particular "False" value).

Task-3032464

Todo
- unit tests JS (many2one, try to move cards, etc)
    - check that no RPC call are made until we open the menu
- JS: order properties in the menu
- multiple groupby with lazy=False