[IMP] mail: Show reorder template response
If a tracked value sent a mail template on creation of a record,
the template was displayed before the original message in the chatter.

-----

Additionally, as the template response needs to be published as a note
it does not look consistent with other messages sent to the user.

Responding to these messages would not notify followers of discussion
which for most models meant nobody was notified of responses by default.

This adds the 'external notifications' message type
which replace system notifications when responses should be discussions.

Introduces a new test model from #91529
'mail.test.ticket' uses an abnormal _track_template override
that sets the composer to 'mass_mail' mode.
This is incompatible with message_type and not the way
that method would typically be used.

Task-2834304