[FIX] web_editor: dispose of the tooltip when closing the bg-overlay
Prior to this commit and in between 15.3 (not incl.) and 16.0, the
tooltip for the background overlay was no longer disposed when creating
a new one. This meant that new tooltips kept being added and could lead
to a very confusing user interface.

This commit dispose of the tooltip when the overlay is disabled.

Steps to reproduce:
- Edit a website and drop a cover snippet
- Toggle the background overlay (target icon)
- Close the overlay (discard, save or click on the editor)
- Change Parallax from fixed to Top to Bottom (this is just so that the
previous tooltip is clearly visible)
- Toggle the background overlay again.
- The old tooltip is still present and a new one is visible.

task-3119958