[IMP] mail: refine company context for layout template
Some mail template use responsible company found on the record
(.user_id.company_id) if company_id is not found on the record. To avoid any
discrepancy between the email body and the company logo in the layout, the same
is now applied to determine the company for the template layout.

Task-2810103