[FIX] base: fix the group warning message on user form view
Before this commit, user form view > access rights mismatch bootstrap
alert > if the group is not linked to any app(category) then `False`
was displayed in the group warning message. while those groups on user
form view are displayed inside the other category.

So in this commit, display the `other` instead of false when group
does not have any category and only display this part of the warning
when the user is in debug mode.

task-3049636