[IMP] sale_timesheet: display SOL field for timesheets items in the portal view of a task
project.task portal form view : The SOL field is displayed in the list of timesheets entries if some of those entries have a different SOL as the one set for the task. The SOL name is made clickable and redirect to the sale order portal view if the user has access to it.

PR: https://github.com/odoo/odoo/pull/103434
task: 3024151