[REF] mail: move composer authorship fields from onchange to computed
PURPOSE

Purpose of this task is to remove the _onchange_template_id method on composer
model. Split it into editable computed stored fields. It gives a better control
of value generation and avoid having to call the onchange when composer is
invoked in code.

SPECIFICATIONS

Improve composer to correctly compute fields ``email_from`` and ``author_id``
in compute method instead of in a an onchange method + default_get support.

Computation is

  * ``email_from`` is directly computed based on template, or fallback on
    given author or current user. This may lead to some change compared to
    previous implementation where email was not synchronized with author. This
    should now be a bit more the case;
  * ``author_id`` is computed based on email_from or current user if not set
    then left unchanged. This should globally be the same behavior as before
    even if it is a bit strange. Currently templates do not allow to tweak
    author field. Moreover changing author based on email_from is not easy
    as email_from could contain inline template expressions;

Task-2088884 (MailComposer: onchange to computed stored fields)