[REF] mail: move composer reply fields from onchange to computed
PURPOSE

Purpose of this task is to remove the _onchange_template_id method on composer
model. Split it into editable computed stored fields. It gives a better control
of value generation and avoid having to call the onchange when composer is
invoked in code.

SPECIFICATIONS

Improve composer to correctly compute fields ``reply_to(_force_new)`` in
compute methods instead of in an onchange method and in default_get.

Computation is

  * ``reply_to`` is directly computed based on template if set. This should not
    change from previous implementation;
  * ``reply_to_force_new`` computation is moved from default_get to a compute
    field based on model. This should not change from previous implementation;

Task-2088884 (MailComposer: onchange to computed stored fields)