[REF] mail: move composer content fields from onchange to computed
Purpose of this commit is to cleanup composer code by gradually removing
the onchange on template, using editable stored computed fields instead.

Improve composer to correctly compute fields ``body`` and ``subject``
in compute methods instead of in a big onchange method and/or in a default_get.

Computation is

Side effects: ??

Task-2088884 (MailComposer: onchange to computed stored fields)