[REF] mail: move composer content fields from onchange to computed
PURPOSE

Purpose of this task is to remove the _onchange_template_id method on composer
model. Split it into editable computed stored fields. It gives a better control
of value generation and avoid having to call the onchange when composer is
invoked in code.

SPECIFICATIONS

Improve composer to correctly compute fields ``body`` and ``subject`` in
compute methods instead of in an onchange method + in default_get.

Computation is based on template: when updating template, update both fields
with template value.

Task-2088884 (MailComposer: onchange to computed stored fields)