[IMP] crm: provide crm_lead values for related res_partner creation
Define information taken from the lead when a partner is automatically created
from it. It avoids the user to enter the information twice.

Technical note: In TestCRMLead.test_message_recipient_partner_auto_creation, we
test that the _message_get_suggested_recipients returns the correct default
values to copy some fields from the lead to the partner to avoid the user to
reenter them. But we don't test specifically that it also work when using mail
composer and template (which generate the partner) because it has already been
tested when introducing the mechanism. Although, we verify that it should
behave the same because we verify that the default values returned by
_related_res_partner_default_create_values, that is used in mail template when
using the composer, are the same as the ones returned by
_message_get_suggested_recipients which is tested here.

Task-3024050