[IMP] crm: use updated frequencies to comupte new proba
This task deals with the following issue:

TO REPRODUCE:

- Create an empty db (no demo data for crm)
- create one lead
- set it won. Probability is set to 100% as expected
- change the stage back to a previous stage (save on each stage change)
- Proba is now 91,67% (1.1/1.2)

ISSUE:

This is because when changing to the previous stage, the frequency table is
updated AFTER the update of the probability value
- First the compute sets the new probability value using the table
- Then in the write, we have the table update

FIX:

The stage_id is removed from the depends of the (automated_)probability,
since it is now called directly in the write. By calling the method
computing probabilities at the end of the write method, we can check
if one of the custom pls fields (or stage_id or team_id, of course) is
changed on certain leads.

As these are the ones changing probabilities, as soon as one of them
changes, we can recompute the probability for this lead.

Task-2673133