[IMP] partner_commission: account manager access rights
This commit intends to give access to account managers to more records of the
partners they manage.

It mainly gives a range of new permissions to the Commission/Own Documents and
Commission/All Documents groups as well as changing stat buttons to include
records where the partner is set as referrer.

Commission Managers have access to sale orders/purchase orders/subscriptions/
invoices with a "referrer" set while Commission Users have access to a range of
documents where the `partner_id` field is an account they manage or the
`referrer_id` field is an account they manage. The children of the managed
partner also count as managed partners.

The groups also give access to Invoicing and Accounting top menu so that members
of the group can easily access account.move records.

Task-2842675