[IMP] mass_mailing: set mailing responsible as author when sent by odoobot
Having odoobot as author of a mailing is not helpful especially if we want to
determine why blacklisted contacts have been included. This solves the problem
by setting the mailing responsible as author instead of odoobot when the
mailing is sent by odoobot (cron task). Indeed, it is a good choice because the
mailing responsible has created the mailing and probably configured it
entirely.

Task-2834862