[IMP] hr_holidays: send time off cancellation notification to manager
Purpose:

User should be abled to delete him time off (time off that needs to be
validated by both, manager and time off officer), if it has not been validated
by officer yet.

The manager that validated the time off, should be notified
if/when the time off is canceled by the user.

Both manager and user should be notified if officer refused time off,
validated by manager.

This commit also fixes traceback on time off form in development mode.

If there is

<field name="A"/>
<label for="A"/>

we get traceback in dev mode.

Instread it should be

<label for="A"/>
<field name="A"/>

task - 2857233