[IMP] mail, various: cleanup usage of mail.composer.mixin
Purpose of this commit is to cleanup usage of mail.composer.mixin.

Email_from fields are not necessary as anyway we always use the current
user's email and field is not present in form view. Override of default_get
to raise about invalid email_from is not necessary as anyway it raises
when posting a message. Giving author_id to message_post is sufficient
as it computes email_from based on author when given.

Add render_model in views, as it is part of the domain given to template_id
field of composer mixin. When missing, also add lang and template_id (coming
from the render mixin).

Use _render_field when possible instead of _render_template as its API
is simpler and uses field definition for rendering parameters.

Task-2710804 (Mail: Clean MailThread Posting API)