[IMP] website: reset a form properly for each submission success type
Before this commit, some behaviors of the forms after being submitted
were problematic:

- with the `On Success` option set on `Redirect`, when going back to a
form after submitting it, the fields were still filled. This also means
that the the previoulsy uploaded files remained, despite the file
blocks not being displayed anymore, which was confusing.

- with the `Show Message` option selected, when going in edit mode when
the message was displayed, the submit button was still "loading", even
after saving.

This commit solves this issues by resetting the form and restoring the
loading effect of the submit button for all the submission `On Success`
actions.

task-2798576