[FIX] website,website_sale: stay on the backend view
Since [the merge of the frontend into the backend] and more precisely
since [this commit], some scenarios make the user switch from the backend
view to the frontend view.

For example:
- Go to /blog from the backend (/@/blog)
- Click on a tag (eg: adventure)

or

- Install ecommerce
- Go to the settings
- Set the "Add to cart" setting to "Go to cart".
- Go to a product page via the backend
- Click on add to cart

=> In both cases, users are redirected to the frontend view, we do not
want that. This commit makes the user stay in the backend. For some
scenarios (like the ones above), we create a fake form and submit it.
The forms have a target attribute that specifies where the form response
should be displayed. This commit set back the default value for the
target attribute so that the response is displayed in the current
context (the iframe when the user is in the backend).

[the merge of the frontend into the backend]: https://github.com/odoo/odoo/commit/31cc10b91dc7762e23b4bde9b945be0c4ce3fe3b
[this commit]: https://github.com/odoo/odoo/commit/2d44f2792dec0b2f205475a22dbedc97e9c54a64

task-3054970