[FIX] account, l10n_X: extend the account tour
This commit add the missing steps (e.g. required fields added by
specific localizations) to the account tour.

Before runing the test that run the tour account_tour, if the
company country is not the same as the fiscal country, try
to find a company with the same country to avoid having issues
when the tour need some demo data by the fiscal country.