[IMP] calendar_sms: allow skipping reminders for responsible
In order to prevent sms spamming the responsible of events, for users
having a lot of calendar events, like hairdressers (through appointment),
we add a field on the calendar.alarm model to implement this option.
As it is meant as a default behavior, we set the default value of
do_not_remind_responsible to True.

This field is only meant for sms alarm_type. Indeed, by default, email
are not issued to their author and if the responsible is set in the
calendar_event, it will be used as author. Notifications are not
concerned by this issue as we consider we want to always receive the
internal reminder.

In order to distinguish between sms alarms with or without responsible
reminder, we add " - Remind Responsible " to the sms calendar alarms
with the option set to True. (This is also useful when alarms are used
in a m2m, like for appointment types or calendar events.)

Task-3032570