[FIX] website: confirmation modal on discard
Before this commit:

When entering edit mode on the website and discarding directly, the editor
considers that there are changes in the DOM and displays a modal with
"Are you sure you want to discard because there are pending changes".

After this commit:

Now, when entering edit mode on the website and discarding it directly will not
display a confirmation modal.

Task-3056463