[FIX] mass_mailing: remove snippets for mobile
When opening the mailing editor from a mobile view the side menu should
be hidden from the user to save space.

Update code to work with the new html_field model and use the existing
destroy method provided by the snippets widget to properly remove
everything related to it (class in the iframe that adds padding for it)

``this.$content`` is not set in the new html_field. As the parent
class sets the correct defaults, the argument is removed from the
last place it was being used.

task - 3062104