[IMP] red invoice smart button if draft inv on SO/PO view
Show red "invoices" smart button on sale/purchase orders if any invoice is still in draft state.
As draft invoices (and downpayments) will be taken into account into the computing of the amount when creating a new invoice for the SO/PO, it is usefull for the user to know that one or more existing invoices are still in draft state.

Task: 2678531