[FIX] web_editor: keep plain text email template padding in conversion
We want some padding on the plain text email template but it was lost in
conversion because the padding was applied to a table with
`border-collapse: collapse` so it was not applied
(see [mdn](https://developer.mozilla.org/en-US/docs/Web/CSS/border-collapse)).
Since we can't change that property (or layouts will be broken), our
remaining option is to wrap the layout table's contents in a `div` and
apply the padding to it instead.

In order to contain the fix and since it's difficult to test every
possible layout, this is done only for the layout of the plain text
template.

task-3062027