[FIX] website: fix call to action button in off-canvas menu
When the off-canvas option of the navbar is activated, the "call to
action" button of the "boxed" header is no longer accessible on mobile.

As this commit is in stable version, the fix is only in Javascript. It
puts the "call to action" button in the navbar when the off-canvas menu
is opened.

task-3063878