[FIX] account-l10n_ch: solve sepa vs swiss QR problems
In Switzerland, adding an extra page containing a QR Bill is mandatory in many cases, mainly when the customer is also from Switzerland (although there are other conditions).

However, activating the option 'QR Codes' in the settings (which is not linked to the QR Bill) can cause problem.

For instance, the Swiss QR code will be repeated multiple times accross the pages, or it will be impossible to bill a foreign customer.

The new behaviour is :

- Swiss user --> swiss customer: don't change the invoice, allow to create a QR Bill
- SEPA option activated, swiss user --> SEPA customer : join the SEPA QR to the invoice
- SEPA option activated, swiss user --> swiss customer : raise error

Also renamed the QR option to lift some confusion.

task-3062570